[android][ios] Add "Details on ..." button to hotels
Add an experimental referral link "Details on KAYAK" to hotels on the map.
KAYAK is a metasearch engine that aggregates hundreds of travel sites
at once to find the best deals on the hotel bookings. The website provides
detailed property listings with photos, reviews, and information on
amenities and prices.

As always, Organic Maps doesn’t collect or share any information about
you and/or your device with KAYAK and/or its partners. As KAYAK is
a traditional website, be aware that they may have tracking on their
website like all travel websites.

All hotel bookings via KAYAK using links from the app will generate
a tiny commission to help fund the development of the project. Thanks for
your support of the project!

Signed-off-by: Roman Tsisyk <roman@tsisyk.com>